Fix #526 to allow `private` modifiers within multi-line `{}` blocks
- Split up `UnitTests.scala` into a `unit/` package with multiple test files, including a new `ParserTests.scala` that tests this new behavior (along with the old)

- Merge `Parsers.unwrapBlock`/`BlockUnwrapper` into `Parsers.split`/`Splitter`, to give us a single source of truth about how to parse a chunk of user input

- Avoid unnecessarily double-splitting input in the `TestRepl` (once outside `processLine`, then again inside `processLine`) since splitting now simultaneously strips wrapping `{...}`s, and is no longer idempotent.

Notably, the old code-flow used to `split` -> `unwrap` -> `split` again, while the new code flow only does a single combined `unwrap-split` operation.

The only reason the old code worked is that the previous allowed-syntax in a REPL command was identical to an arbitrary Scala expression. In the case of `{}`-wrapped input, the old code thus did it's first `split` without any effect, but without any error either, while the second `split` after the `unwrap` did the real work.

 It is no longer the case that REPL commands syntax is the same as Scala expressions, as `{ private val x = 1 }` is not a valid Scala block. Thus we have to be more careful to ensure we are only `unwrap-split`-ing once on each input.

Misc Fixes:

- Convert some integration tests from `==` to `.contains`, to allow for some unexpected "Compiling foo..." messages to appear in the stderr without needing to assert on them

- Fix logging in `TestRepl` to add back missing newlines between log messages

- Rename `commandLinePredef` back to `predef` and `mainPredef` to avoid unnecessary API breakage
